refactor: Polish clockify module
