+++ src/Clockify/Models/TaskDo.cs
@@ -2,7 +2,7 @@
 {
     public class TaskDo
     {
-        public string Id { get; set; }
+        public string? Id { get; set; }
 
         public string Name { get; set; }
