+++ src/Clockify/Reports/IReportSummaryService.cs
@@ -1,5 +1,4 @@
 ﻿using System.Threading.Tasks;
-using Bot.Models;
 using Bot.States;
 
 namespace Bot.Clockify.Reports
